feat: implemented abstract base key handle and deriviates
