+++ src/crypto-layer/index.ts
@@ -2,5 +2,5 @@ export * from "./CryptoDerivationHandle";
 export * from "./CryptoLayerConfig";
 export * from "./CryptoLayerProviders";
 export * from "./CryptoLayerUtils";
-export * from "./encryption/CryptoEncryption";
+export * from "./encryption/CryptoEncryptionHandle";
 export * from "./encryption/CryptoSecretKeyHandle";
